TIP-885: move locale provider
