+++ src/Akeneo/Platform/Bundle/UIBundle/Resources/config/locale_provider.yml
@@ -1,11 +1,11 @@
 parameters:
-    pim_localization.provider.ui_locale.class:          Pim\Bundle\LocalizationBundle\Provider\UiLocaleProvider
+    pim_localization.provider.ui_locale.class:
     pim_localization.provider.ui_locale.min_percentage: 0.8
     pim_localization.provider.ui_locale.locale_codes:   ['en_US', 'ca_ES', 'da_DK', 'de_DE', 'es_ES', 'fi_FI', 'fr_FR', 'hr_HR', 'it_IT', 'ja_JP', 'nl_NL', 'pl_PL', 'pt_BR', 'pt_PT', 'ru_RU', 'sv_SE', 'tl_PH', 'zh_CN', 'sv_SE']
 
 services:
     pim_localization.provider.ui_locale:
-        class: '%pim_localization.provider.ui_locale.class%'
+        class: 'Akeneo\Platform\Bundle\UIBundle\UiLocaleProvider'
         arguments:
             - '@translator'
             - '%pim_localization.provider.ui_locale.min_percentage%'
